dev feature: mongodb integration: dev storage methods: in case personal computer crashed

+++ runtime/src/main/java/com/alibaba/smart/framework/engine/instance/storage/TaskInstanceStorage.java
@@ -29,8 +29,8 @@ public interface TaskInstanceStorage {
     List<TaskInstance> findTaskList(TaskInstanceQueryParam taskInstanceQueryParam,
                                     ProcessEngineConfiguration processEngineConfiguration);
 
-    Integer count(TaskInstanceQueryParam taskInstanceQueryParam,
-                  ProcessEngineConfiguration processEngineConfiguration);
+    Long count(TaskInstanceQueryParam taskInstanceQueryParam,
+               ProcessEngineConfiguration processEngineConfiguration);
 
     TaskInstance insert(TaskInstance taskInstance,
                         ProcessEngineConfiguration processEngineConfiguration);
